requirements.txt
setuptools~=58.3.0
vdf~=3.4
orjson~=3.6.4
requests~=2.26.0
pyxdg~=0.27
fusepy~=3.0.1